+++ src/DownloaderParallel.h
@@ -1,12 +1,11 @@
 /**
- * DownloaderParallel.h
- *
- * A simple C++ wrapper for the libcurl multiple API.
- * Parallel curl handles execution
- * using synchronous I/O multiplexing: select() call. 
+ * @file DownloaderParallel.h
+ * @brief A several files parallel downloader
  */
+
 #ifndef DOWNLOADERPARALLEL_H
 #define DOWNLOADERPARALLEL_H
+
 #include <string>
 #include <vector>
 #include <curl/curl.h>
@@ -14,9 +13,17 @@
 #include "ICurlEasyDownloader.h"
 
 /**
- * A threadsafe simple libcURL-multi based downloader:
- * execute all CURL* handles from ICurlEasyDownloader
- * using synchronous I/O multiplexing: select() call.
+ * @brief Download several files in parallel using libcurl multi API
+ * @version 1.0.0
+ * @author chakaponden (itransition.com)
+ * @date 4 November 2017
+ * @copyright MIT License
+ * @details A threadsafe simple libcURL-multi based downloader:
+ *          parallel easy curl handles execution
+ *          using synchronous I/O multiplexing: select() call.
+ * @warning All easy curl handles (ICurlEasyDownloader downloaders) data
+ *          must by ready and only wait for curl_easy_perform() call
+ * @todo Implementation with epoll() system call
  */
 class DownloaderParallel
 {
@@ -24,17 +31,35 @@ public:
     DownloaderParallel();
     ~DownloaderParallel();
 
-    void AddDownloader(const ICurlEasyDownloader* downloader);
-    void RemoveDownloader(const ICurlEasyDownloader* downloader);
-
     /**
-     * Execute all CURL* handles from ICurlEasyDownloader
-     * @return The execution result code
+     * @brief Append ICurlEasyDownloader derivative downloader
+     *        to execution queue for parallel downloading
+     * @param[in] downloader ICurlEasyDownloader derivative downloader
+     */
+    void AddDownloader(const ICurlEasyDownloader* downloader);
+    /**
+     * @brief Remove ICurlEasyDownloader derivative downloader
+     *        from execution queue for parallel downloading
+     * @param[in] downloader ICurlEasyDownloader derivative downloader
+     */
+    void RemoveDownloader(const ICurlEasyDownloader* downloader);
+    /**
+     * @brief Process parallel download for all ICurlEasyDownloader
+     *        derivative downloaders in execution queue
+     * @return The download execution result code
+     * @return 0:   successfully did nothing
+     * @return 1:   curl_multi_perform() failed
+     * @return 2:   curl multi is not initialized
      */
     int Download();
+    /**
+     * @brief Same as Download()
+     */
     int operator()();
 
 private:
+    /// LibcURL multi handle provides parallel execution.
     CURLM* _curlMultiHandle;
 };
+
 #endif  /* DOWNLOADERPARALLEL_H */